requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
google-generativeai==0.1.0  # Replace with the actual version if known
Pillow==8.4.0
pytesseract==0.3.8